% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_block.R
\name{bootstrap_block}
\alias{bootstrap_block}
\title{Implements a block bootstrap}
\usage{
bootstrap_block(data, k = 14)
}
\arguments{
\item{data}{Data frame of raw data detrended if necessary. First column should be of the \code{Date}.}

\item{k}{Numeric vector of length one specifying the block length. Default is \code{14}.}
}
\value{
Dataframe containing th eblock bootstap sample.
}
\description{
This function performs block bootstrapping on a given dataset.
}
\examples{
#Boostrap the detrended data at site S-22
boot_df = bootstrap_block(S22.Detrend.df[,2:3])
#Calculate the mean of the drivers in bootstrapped sample
apply(boot_df,2,function(x) mean(x,na.rm=T))
}
