% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Sample.R
\name{C_Sample}
\alias{C_Sample}
\title{Implements bootstrap procedure in Serinaldi and Kilsby (2013)}
\usage{
C_Sample(Data, Cluster_Max, D, Start, End, Xp)
}
\arguments{
\item{Data}{Vector of the rainfall time series.}

\item{Cluster_Max}{Vector of the index of \code{Data} containing the cluster maximum. If declustering is carried out using \code{Decluster_SW()} set equal to \code{$EventsMax} output.}

\item{D}{Numeric vector of the duration of the cluster maximum events.}

\item{Start}{Numeric vector of the index of \code{Data} where each cluster maximum event begins.}

\item{End}{Numeric vector of the index of \code{Data} where each cluster maximum event ends.}

\item{Xp}{Numeric vector of simulated peaks. To implement the method exactly as in Serinaldi and Kilsby (2013), set equal to a sample (taken with replacement) of the observed cluster maximum (peaks).}
}
\value{
A data frame with the following columns:  \itemize{
\item \code{Xp} Simulated event peaks i.e. input \code{Xp}.
\item \code{D} Duration sampled from the duration vector \code{D} for each simulated event.
\item \code{Samp} Index of the cluster maximum event, sampled conditionally on \code{D}, that provides non-peak rainfall depths.
\item \code{V} Volume of simulated events.
\item \code{Vn} Net volume of simulated events.
\item \code{I} Intensity of simulated events.
\item \code{In} Net intensity of simulated events.
\item \code{Start} Index of \code{Data} where the sampled (\code{Samp}) event begins.
\item \code{End} Index of \code{Data} where the sampled (\code{Samp}) event ends.
}
}
\description{
Implements the conditional bootstrap procedure outlined in Serinaldi and Kilsby (2013) to generate non-peak rainfall totals for a simulated peak. The function also calculates hyetograph properties including net characteristics.
}
\examples{
Set very small rainfall measurements to zero.
#Assumed to be the result of uncertainty in measuring equipment.
S13_Rainfall$Rainfall[which(S13_Rainfall$Rainfall<0.01)] = 0
#Find NAs in rainfall series
z = which(is.na(S13_Rainfall$Rainfall)==T)
#Temporarily set NAs to zero
S13_Rainfall$Rainfall[z] = 0
#Find times where there is 6-hours of no rainfall
no.rain = rep(NA,length(S13_Rainfall$Rainfall))
for(i in 6:length(S13_Rainfall$Rainfall)){
  no.rain[i] = ifelse(sum(S13_Rainfall$Rainfall[(i-5):i])==0,i,NA)
}
#Remove NAs from results vector as these correspond to times where there is
#rainfall at certain points in the 6 hour period.
no.rain = na.omit(no.rain)
#Reset missing values in the rainfall record back to NA
S13_Rainfall$Rainfall[z] = NA
#Find the start and end times of the 500 events.
start = rep(NA,length(S13.Rainfall.Declust$EventsMax))
end = rep(NA,length(S13.Rainfall.Declust$EventsMax))
for(i in 1:length(S13.Rainfall.Declust$EventsMax)){
 start[i] = max(no.rain[which(no.rain<S13.Rainfall.Declust$EventsMax[i])])
 end[i] = min(no.rain[which(no.rain>S13.Rainfall.Declust$EventsMax[i])])
}
start = start + 1
end = end - 6
d = end - start + 1 #Duration
#Simulate some peaks by sampling observed peaks with replacement
#I.e., applying the method exactly as in Serinaldi and Kilsby (2013)
sim.peak = sample(S13.Rainfall.Declust$EventsMax,size=500,replace=TRUE)
#Derive the hyetographs
S13.oswl.sample = C_Sample(Data=S13_Rainfall$Rainfall,
                           Cluster_Max=S13.Rainfall.Declust$EventsMax,
                           D=d,Start=start,End=end,
                           Xp=sim.peak)
}
\seealso{
\code{\link{Decluster}} \code{\link{Time_Series_Plot}} \code{\link{WL_Curve}}
}
