% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Con_Sampling_2D.R
\name{Con_Sampling_2D}
\alias{Con_Sampling_2D}
\title{Conditionally sampling a two dimensional dataset}
\usage{
Con_Sampling_2D(Data_Detrend, Data_Declust, Con_Variable, Thres = 0.97)
}
\arguments{
\item{Data_Detrend}{Dataframe containing two at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned \code{NA}. First object may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}

\item{Data_Declust}{Dataframe containing two (independently) declustered at least partially concurrent time series. Time steps must be equally spaced, with missing values assigned \code{NA}. Columns must be in the same order as in \code{Data_Detrend}. First object may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}

\item{Con_Variable}{Column number (1 or 2) or the column name of the conditioning variable. Default is \code{1}.}

\item{Thres}{Threshold, as a quantile of the observations of the conditioning variable. Default is \code{0.97}.}
}
\value{
List comprising the specified \code{Threshold} as the quanitle of the conditioning variable above which declustered excesses are paired with co-occurances of the other varable, the resulting two dimensional sample \code{data} and \code{name} of the conditioning variable.
}
\description{
Creates a dataframe where the declustered excesses of a (conditioning) variable are paired with co-occurances of another varable.
}
\examples{
S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
             Con_Variable="Rainfall",Thres=0.97)
}
