% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditional_RP_2D.R
\name{Conditional_RP_2D}
\alias{Conditional_RP_2D}
\title{Calculates joint and conditional return periods}
\usage{
Conditional_RP_2D(
  Data,
  Data_Con1,
  Data_Con2,
  Thres1,
  Thres2,
  Copula_Family1,
  Copula_Family2,
  Marginal_Dist1,
  Marginal_Dist2,
  Con1 = "Rainfall",
  Con2 = "OsWL",
  mu = 365.25,
  Con_Var,
  RP_Con,
  RP_Non_Con,
  Var1,
  Var2,
  x_lab = "Rainfall (mm)",
  y_lab = "O-sWL (mNGVD 29)",
  x_lim_min = NA,
  x_lim_max = NA,
  y_lim_min = NA,
  y_lim_max = NA,
  N
)
}
\arguments{
\item{Data}{Data frame of dimension \code{nx2} containing two co-occurring time series of length \code{n}.}

\item{Data_Con1}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column.}

\item{Data_Con2}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column. Can be obtained using the \code{Con_Sampling_2D} function.}

\item{Thres1}{Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in Data_Con1.}

\item{Thres2}{Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in Data_Con2.}

\item{Copula_Family1}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con1} dataset.}

\item{Copula_Family2}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con2} dataset. Best fitting of 40 copulas can be found using the \code{Copula_Threshold_2D} function.}

\item{Marginal_Dist1}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con1}.}

\item{Marginal_Dist2}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con2}.}

\item{Con1}{Character vector of length one specifying the name of variable in the first column of \code{Data}.}

\item{Con2}{Character vector of length one specifying the name of variable in the second column of \code{Data}.}

\item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in \code{Data}. Default is \code{365.25}, daily data.}

\item{Con_Var}{Character vector of length one specifying the (column) name of the conditioning variable.}

\item{RP_Con}{Numeric vector of length one specifying the return period of the conditioning variable \code{Con_Var}.}

\item{RP_Non_Con}{Numeric vector of length one specifying the return period of the non-conditioning variable.}

\item{x_lab}{Character vector specifying the x-axis label.}

\item{y_lab}{Character vector specifying the y-axis label.}

\item{x_lim_min}{Numeric vector of length one specifying x-axis minimum. Default is \code{NA}.}

\item{x_lim_max}{Numeric vector of length one specifying x-axis maximum. Default is \code{NA}.}

\item{y_lim_min}{Numeric vector of length one specifying y-axis minimum. Default is \code{NA}.}

\item{y_lim_max}{Numeric vector of length one specifying y-axis maximum. Default is \code{NA}.}

\item{N}{Numeric vector of length one specifying the size of the sample from the fitted joint distributions used to estimate the density along an isoline. Samples are collected from the two joint distribution with proportions consistent with the total number of extreme events conditioned on each variable. Default is \code{10^6}}
}
\value{
Console output: \itemize{
\item Con_Var
Name of the conditioning variable
\item RP_Var1
Return period of variable Con1 i.e., variable in second column of \code{Data}
\item RP_Var2
Return period of variable Con2 i.e., variable in third column of \code{Data}
\item Var1
Value of Con1 at the return period of interest i.e. RP_Var1
\item Var2
Value of Con2 at the return period of interest i.e. RP_Var2
\item RP_Full_Dependence
Joint return period of the (Var1,Var2) event under full dependence
\item RP_Independence
Joint return period of the (Var1,Var2) event under independence
\item RP_Copula
Joint return period of the (Var1,Var2) event according to the two sided conditional sampling - copula theory approach
\item Prob
Probability associated with \code{RP_Copula}
\item N_Excess
Number of realizations of the \code{Con_Var} above \code{RP_Con}-year return period value
\item Non_Con_Var_X
Values of the non-conditioned variable of the (conditional) Cummulative Distribution Function (CDF) i.e. x-axis of bottom left plot
\item Con_Prob
\code{Con_Prob} CDF of the non-conditioned variable given the return period of \code{Con_Var} exceeds \code{RP_Con}
\item Con_Prob_Est
Probability the non-conditioned variable is less than or equal to \code{RP_Non_Con} given the return period of \code{Con_Var} exceeds \code{RP_Con}
}
Graphical output: \itemize{
\item Top left: Sample conditioned on Con1 (red crosses) and Con2 (blue circles). Black dot is the event with a marginal return period of the conditioned variable \code{Var_Con} and non-conditioned variable equal to \code{RP_Con} and \code{RP_Non_Con}, respectively. The joint return period of the event using the conditional sampling - copula theory approach and under the assumptions of full dependence and independence between the variables are printed.
\item Top right: Sample conditioned on Con1 (red crosses) and Con2 (blue circles). Only the region where Con_Var exceeds RP_Con is visible. This is the region for which the conditional distribution (of the non-conditioned variable given Con_Var exceeds RP_Con) and in turn conditional return periods are calculated.
\item Bottom left: Conditional Cumulative Distribution Function (CDF) of the non-conditioned variable given the marginal return period of the conditioned variable \code{Var_Con} exceeds \code{RP_Con} years i.e. the points visible in the top right plot.
\item Bottom right: Conditional return period of the non-conditioned variable given the conditioned variable \code{Var_Con} has a return period longer than \code{RP_Con}.
}
}
\description{
Univariate return period events are obtained from the GPDs to be consistent with the isolines produced by the \code{Design_Event_2D} function.
To find the conditional probabilities a large number of realizations are simulated from the copulas fit to the conditioned samples, in proportion with the sizes of the conditional samples. The realizations are transformed to the original scale and the relevant probabilities estimated empirically.
}
\examples{
#Under a 10yr (or greater) rainfall event condition, what is the joint probability that a 10yr
#O-sWLevent occurs simultaneously?  What is the cumulative probability of events with the
#frequency equal to or less than a 10yr O-sWL event?
Conditional_RP_2D(Data=S22.Detrend.df,
                  Data_Con1=con.sample.Rainfall$Data, Data_Con2=con.sample.OsWL$Data,
                  Thres1=0.98, Thres2=0.98,
                  Copula_Family1=cop.Rainfall,Copula_Family2=cop.OsWL,
                  Marginal_Dist1="Logis", Marginal_Dist2="BS",
                  Con1 = "Rainfall", Con2 = "OsWL",
                  mu = 365.25,
                  Con_Var="Rainfall",
                  RP_Con=10, RP_Non_Con=10,
                  x_lab = "Rainfall (Inches)", y_lab = "O-sWL (ft NGVD 29)",
                  y_lim_max = 10,
                  N=10^7)
}
\seealso{
\code{\link{Design_Event_2D}}
}
