% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditional_RP_2D_Equal.R
\name{Conditional_RP_2D_Equal}
\alias{Conditional_RP_2D_Equal}
\title{Calculates joint and conditional return periods}
\usage{
Conditional_RP_2D_Equal(
  Data,
  Data_Con1,
  Data_Con2,
  u1,
  u2,
  Thres1 = NA,
  Thres2 = NA,
  Copula_Family1,
  Copula_Family2,
  Marginal_Dist1,
  Marginal_Dist2,
  Con1 = "Rainfall",
  Con2 = "OsWL",
  mu = 365.25,
  Con_Var,
  RP_Con,
  RP_Non_Con,
  Width = 0.1,
  x_lab = "Rainfall (mm)",
  y_lab = "O-sWL (mNGVD 29)",
  x_lim_min = NA,
  x_lim_max = NA,
  y_lim_min = NA,
  y_lim_max = NA,
  DecP = 2,
  N
)
}
\arguments{
\item{Data}{Data frame of dimension \code{nx2} containing two co-occurring time series of length \code{n}.}

\item{Data_Con1}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column.}

\item{Data_Con2}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column. Can be obtained using the \code{Con_Sampling_2D} function.}

\item{u1}{Numeric vector of length one specifying the (quantile) threshold above which the variable in the first column was sampled in \code{Data_Con1}.}

\item{u2}{Numeric vector of length one specifying the (quantile) threshold above which the variable in the second column was sampled in \code{Data_Con2}.}

\item{Thres1}{Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in \code{Data_Con1}. Only one of \code{u1} and \code{Thres1} should be supplied. Default is \code{NA}.}

\item{Thres2}{Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in \code{Data_Con2}. Only one of \code{u2} and \code{Thres2} should be supplied. Default is \code{NA}.}

\item{Copula_Family1}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con1} dataset.}

\item{Copula_Family2}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con2} dataset. Best fitting of 40 copulas can be found using the \code{Copula_Threshold_2D} function.}

\item{Marginal_Dist1}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con1}.}

\item{Marginal_Dist2}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con2}.}

\item{Con1}{Character vector of length one specifying the name of variable in the first column of \code{Data}.}

\item{Con2}{Character vector of length one specifying the name of variable in the second column of \code{Data}.}

\item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in \code{Data}. Default is \code{365.25}, daily data.}

\item{Con_Var}{Character vector of length one specifying the (column) name of the conditioning variable.}

\item{RP_Con}{Numeric vector of length one specifying the return period of the conditioning variable \code{Con_Var}.}

\item{RP_Non_Con}{Numeric vector of length one specifying the return period of the non-conditioning variable.}

\item{Width}{Numeric vector of length one specifying the distance above and below the \code{RP_Con} event of \code{Con_Var} the simulated events are used to estimate the conditional probability.}

\item{x_lab}{Character vector specifying the x-axis label.}

\item{y_lab}{Character vector specifying the y-axis label.}

\item{x_lim_min}{Numeric vector of length one specifying x-axis minimum. Default is \code{NA}.}

\item{x_lim_max}{Numeric vector of length one specifying x-axis maximum. Default is \code{NA}.}

\item{y_lim_min}{Numeric vector of length one specifying y-axis minimum. Default is \code{NA}.}

\item{y_lim_max}{Numeric vector of length one specifying y-axis maximum. Default is \code{NA}.}

\item{DecP}{Numeric vector of length one specifying the number of decimal places to round the data in the conditional samples to in order to identify observations in both conditional samples. Default is \code{2}.}

\item{N}{Numeric vector of length one specifying the size of the sample from the fitted joint distributions used to estimate the density along an isoline. Samples are collected from the two joint distribution with proportions consistent with the total number of extreme events conditioned on each variable. Default is \code{10^6}}
}
\value{
Console output: \itemize{
\item Con_Var
Name of the conditioning variable
\item RP_Var1
Return period of variable Con1 i.e., variable in second column of \code{Data}
\item RP_Var2
Return period of variable Con2 i.e., variable in third column of \code{Data}
\item Var1
Value of Con1 at the return period of interest
\item Var2
Value of Con2 at the return period of interest
\item RP_Full_Dependence
Joint return period of the (Var1,Var2) event under full dependence
\item RP_Independence
Joint return period of the (Var1,Var2) event under independence
\item RP_Copula
Joint return period of the (Var1,Var2) event according to the two sided conditional sampling - copula theory approach
\item Prob
Probability associated with \code{RP_Copula}
\item N_Sub_Sample
Number of realizations of the \code{Con_Var} within +/- width of the value of \code{Con_Var} with return period \code{}.
\item Non_Con_Var_X
Values of the non-conditioned variable of the (conditional) Cummulative Distribution Function (CDF) i.e. x-axis of bottom left plot
\item Con_Prob
\code{Con_Prob} CDF of the non-conditioned variable given the return period of \code{Con_Var} equals \code{RP_Con}
\item Con_Prob_Est
Probability the non-conditioned variable is less than or equal to \code{RP_Non_Con} given the return period of \code{Con_Var} equals \code{RP_Con}
}
Graphical output: \itemize{
\item Top Left: Sample conditioned on rainfall (red crosses) and O-sWL (blue circles). Black dot is the event with a marginal return period of the conditioned variable \code{Var_Con} and non-conditioned variable equal to \code{RP_Con} and \code{RP_Non_Con}, respectively. The joint return period of the event using the conditional sampling - copula theory approach and under the assumptions of full dependence and independence between the variables are printed.
\item Top Right: Sample used to estimate the joint return period of the event of interest. Black dots denote the \code{N_Excess} sized subset of the sample where the marginal return period of the conditioned variable \code{Var_Con} exceeds \code{RP_Con} (years). The subset is used to estimate the conditional probabilities in part two of the question.
\item Bottom Left: Conditional Cumulative Distribution Function (CDF) of the non-conditioned variable given the marginal return period of the conditioned variable \code{Var_Con} exceeds \code{RP_Con} years i.e. the black dots in the top right plot.
\item Bottom Right: Conditional return period of the non-conditioned variable given the conditioned variable \code{Var_Con} has a return period longer than \code{RP_Con}.
}
}
\description{
A large number of realizations are simulated from the copulas fit to the conditioned samples, in proportion with the sizes of the conditional samples. The realization are transformed to the original scale and the relevant probabilities estimated empirically. The conditional probabilities return period of the conditioning variable equals
}
\examples{
#Under a 10yr rainfall event condition, what is the joint probability that a 10yr surge (O-sWL)
#event occurs simultaneously?  What is the cumulative probability of events with the frequency
#equal to or less than a 10yr surge event?
#' #Conditional samples
con.sample.Rainfall<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                     Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                                     Con_Variable="Rainfall",u=0.98)
con.sample.OsWL<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                 Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                                 Con_Variable="OsWL",u=0.98)
#Add some noise to rainfall to aid distribution fitting in Conditional_RP_2D function
con.sample.OsWL$Data$Rainfall<-con.sample.OsWL$Data$Rainfall+
                                runif(length(con.sample.OsWL$Data$Rainfall),0.001,0.01)
#Find the best fitting copula
cop.Rainfall <- Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                    Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                                     u1=0.98, u2=NA, PLOT=FALSE)$Copula_Family_Var1
cop.OsWL<- Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                               Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                               u1=NA, u2=0.98, PLOT=FALSE)$Copula_Family_Var2
#Calculate conditional probabilities
Conditional_RP_2D_Equal(Data=S22.Detrend.df,
                        Data_Con1=con.sample.Rainfall$Data, Data_Con2=con.sample.OsWL$Data,
                        u1=0.98, u2=0.98,
                        Copula_Family1=cop.Rainfall,Copula_Family2=cop.OsWL,
                        Marginal_Dist1="Logis", Marginal_Dist2="Twe",
                        Con1 = "Rainfall", Con2 = "OsWL",
                        mu = 365.25,
                        Con_Var="Rainfall",
                        RP_Con=10, RP_Non_Con=10,
                        x_lab = "Rainfall (Inches)", y_lab = "O-sWL (ft NGVD 29)",
                        y_lim_max = 10,
                        N=10^5)
}
\seealso{
\code{\link{Design_Event_2D}} \code{\link{Conditional_RP_2D}}
}
