% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooley19.R
\name{Cooley19}
\alias{Cooley19}
\title{Derives bivariate isolines using the non-parametric approach of Cooley et al. (2019).}
\usage{
Cooley19(
  Data,
  Migpd,
  p.base = 0.01,
  p.proj = 0.001,
  u = 0.95,
  PLOT = FALSE,
  x_lim_min_T = NA,
  x_lim_max_T = NA,
  y_lim_min_T = NA,
  y_lim_max_T = NA,
  x_lim_min = NA,
  x_lim_max = NA,
  y_lim_min = NA,
  y_lim_max = NA
)
}
\arguments{
\item{Data}{Data frame consisting of two columns.}

\item{Migpd}{An \code{Migpd} object, containing the generalized  Pareto models fitted (independently) to the variables comprising the columns of \code{Data}.}

\item{p.base}{Numeric vector of length one specifying the exceedance probability of the base isoline. Default is \code{0.01}.}

\item{p.proj}{Numeric vector of length one specifying the exceedance probability of the projected isoline. Default is \code{0.001}.}

\item{u}{Numeric vector of length one specifying the quantile at which to estimate the asymptotic nature of the data i.e. chi and chibar. Default is \code{0.95}.}

\item{PLOT}{Logical; indicating whether to plot the base and projected isolines on the original and transformed scale. Default is \code{FALSE}.}

\item{x_lim_min_T}{Numeric vector of length one specifying the lower x-axis limit of the transformed scale plot. Default is \code{NA}.}

\item{x_lim_max_T}{Numeric vector of length one specifying the upper x-axis limit of the transformed scale plot. Default is \code{NA}.}

\item{y_lim_min_T}{Numeric vector of length one specifying the lower y-axis limit of the transformed scale plot. Default is \code{NA}.}

\item{y_lim_max_T}{Numeric vector of length one specifying the upper y-axis limit of the transformed scale plot. Default is \code{NA}.}

\item{x_lim_min}{Numeric vector of length one specifying the lower x-axis limit of the plot on the original scale. Default is \code{NA}.}

\item{x_lim_max}{Numeric vector of length one specifying the upper x-axis limit of the plot on the original scale. Default is \code{NA}.}

\item{y_lim_min}{Numeric vector of length one specifying the lower y-axis limit of the plot on the original scale. Default is \code{NA}.}

\item{y_lim_max}{Numeric vector of length one specifying the lower y-axis limit of the plot on the original scale. Default is \code{NA}.}
}
\value{
List comprising a description of the type of (asymptoptic) dependence \code{Asym}, the values the extremal dependence measures \code{Chi} and \code{n.bar}, exceedance probabilities of the base \code{p.base} and projected \code{p.proj} isolines, as well as the points on the base \code{I.base} and projected \code{I.proj} isolines.
}
\description{
The Cooley et al. (2019) method exploits bivariate regular variation and kernel density estimation to generate isolines of bivariate exceedance probabilities. The function utilizes the \code{ks} and \code{texmex} packages, and works for both asymptotic dependence and independence.
}
\examples{
S20.GPD<-Migpd_Fit(Data=S20.Detrend.Declustered.df[,c(3,4)],
                   Data_Full=S20.Detrend.Declustered.df[,c(3,4)],
                   mqu =c(0.99,0.99))
Cooley19(Data=na.omit(S20.Detrend.df[,3:4]),Migpd=S20.GPD,
p.base=0.01,p.proj=0.001,PLOT=TRUE,x_lim_max_T=500,y_lim_max_T=500)
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Decluster}} \code{\link{GPD_Fit}} \code{\link{Migpd_Fit}}
}
