% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula_Threshold_2D.R
\name{Copula_Threshold_2D}
\alias{Copula_Threshold_2D}
\title{Copula Selection With Threshold 2D - Fit}
\usage{
Copula_Threshold_2D(Data_Detrend, Data_Declust, Thres = seq(0.9, 0.99,
  0.01), x_lim_min = min(Thres), x_lim_max = max(Thres),
  y_lim_min = -1, y_lim_max = 1, Upper = 0, Lower = 0,
  GAP = 0.05)
}
\arguments{
\item{Data_Detrend}{Dataframe containing two at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned \code{NA}.}

\item{Data_Declust}{Dataframe containing two (independently) declustered at least partially concurrent time series. Time steps must be equally spaced, with missing values assigned \code{NA}.}

\item{Thres}{A single or sequence of thresholds, given as a quantile of the observations of the conditioning variable. Default, sequence from \code{0.9} to \code{0.99} at intervals of \code{0.01}.}

\item{x_lim_min}{Numeric vector of length one specifying x-axis minimum. Default is the maximum argument in \code{Thres}.}

\item{x_lim_max}{Numeric vector of length one specifying x-axis maximum. Default is the minimum argument in \code{Thres}.}

\item{y_lim_min}{Numeric vector of length one specifying y-axis minimum. Default \code{-1.0}.}

\item{y_lim_max}{Numeric vector of length one specifying y-axis maximum. Default \code{1.0}.}

\item{Upper}{Numeric vector specifying the element number of the \code{Thres} argument for which the copula family name label to appear above the correponding point on the Kendall's tau coefficent vs threshold plot, when conditioin on the variable in column 1. Default is \code{0}.}

\item{Lower}{Numeric vector specifying the element number of the \code{Thres} argument for which the copula family name label to appear below the correponding point on the Kendall's tau coefficent vs threshold plot, when conditioin on the variable in column 2. Default is \code{0}.}

\item{GAP}{Numeric vector of length one specifying the distance above or below the copula family name label appears the correponding point on the Kendall's tau coefficent vs threshold plot. Default is \code{0.05}.}
}
\value{
List comprising: \itemize{
\item \code{Kendalls_Tau_Var1} Kendalls tau of a sample
\item \code{p_value_Var1}
p-value when testing the null hypiothesis \code{H_0} i.e. that there is no correlation between the variables
\item \code{N_Var1}
size of the dataset
\item \code{Copula_Family_Var1}
best fitting copula for the specified thresholds
}
when the dataset is conditioned on the variable in column 1.
Anologous vector \code{Kendalls_Tau_Var2},\code{p_value_Var2}, \code{N_Var2} and \code{Copula_Family_Var2} for the specified thresholds when the dataset is conditioned on the variable in column 2.
}
\description{
Declustered excesses of a (conditioning) variable are paired with co-occurances of the other varable before the best fitting bivariate copula is selected, using \code{BiCopSelect} function in the \code{VineCopula} package, for a single or range of thresholds. The procedure is automatically repeated with the variables switched.
}
\examples{
Copula_Threshold_2D(Data_Detrend=S28_Detrend,Data_Declust=S28_Detrend_Declustered,y_lim=c(-0.075,0.25),Upper=c(2,9),Lower=c(2,10))
}
