% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decluster_SW.R
\name{Decluster_SW}
\alias{Decluster_SW}
\title{Declusters a time series using a storm window approach}
\usage{
Decluster_SW(Data, Window_Width)
}
\arguments{
\item{Data}{Data frame containing two columns. In column: \itemize{
\item \code{1} A \code{"Date"} object of equally spaced discrete time steps.
\item \code{2} Numeric vector containing corresponding time series values.
}}

\item{Window_Width}{Numeric vector of length one specifying the width, in days, of the window used to ensure events are independent.}
}
\value{
List comprising vectors containing the original time series \code{Detrended}, independent (declustered) events \code{Declustered} and the elements of the original series containing the declustered events \code{EventID}.
}
\description{
Find peaks with a moving window. The code is based on the \code{IDEVENT} function provided by Sebastian Solari.
}
\examples{
#Declustering the O-sWL at site S22 using a 3-day window.
v<-Decluster_SW(Data=S22.Detrend.df[,c(1:2)],Window_Width=7)
plot(as.Date(S22.Detrend.df$Date),S22.Detrend.df$Rainfall,pch=16)
points(as.Date(S22.Detrend.df$Date)[v$EventID],v$Event,col=2,pch=16)
}
