% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design_Event_2D.R
\name{Design_Event_2D}
\alias{Design_Event_2D}
\title{Derives a single or ensemble of bivariate design events}
\usage{
Design_Event_2D(Data, Data_Con1, Data_Con2, Thres1, Thres2, Copula_Family1,
  Copula_Family2, Marginal_Dist1, Marginal_Dist2, Con1 = "Rainfall",
  Con2 = "OsWL", mu = 365.25, RP, x_lab = "Rainfall (mm)",
  y_lab = "O-sWL (mNGVD 29)", x_lim_min = NA, x_lim_max = NA,
  y_lim_min = NA, y_lim_max = NA, N, N_Ensemble)
}
\arguments{
\item{Data}{Dataframe of dimension \code{nx2} containing two co-occurring time series of length \code{n}.}

\item{Data_Con1}{Dataframe containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column.}

\item{Data_Con2}{Dataframe containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column. Can be obtained using the \code{Con_Sampling_2D} function.}

\item{Thres1}{Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in Data_Con1.}

\item{Thres2}{Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in Data_Con2.}

\item{Copula_Family1}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con1} dataset.}

\item{Copula_Family2}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con2} dataset. Best fitting of 40 copulas can be found using the \code{Copula_Threshold_2D} function.}

\item{Marginal_Dist1}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable.}

\item{Marginal_Dist2}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable. Best fitting among two truncted distributions or eight truncated distributions can be found using the functions.}

\item{Con2}{Character vector of length one specifying the name of variable in the first column of \code{Data}.}

\item{mu}{Numeric vector of length one specifying the (average) number of events per year. Default is \code{365.25}, daily data.}

\item{RP}{Numeric vector of length one specifying the return period of interest.}

\item{x_lab}{Charactor vector specifying the x-axis label.}

\item{y_lab}{Charactor vector specifying the y-axis label.}

\item{x_lim_min}{Numeric vector of length one specifying x-axis minimum. Default is \code{NA}.}

\item{x_lim_max}{Numeric vector of length one specifying x-axis maximum. Default is \code{NA}.}

\item{y_lim_min}{Numeric vector of length one specifying y-axis minimum. Default is \code{NA}.}

\item{y_lim_max}{Numeric vector of length one specifying y-axis maximum. Default is \code{NA}.}

\item{N}{Numeric vector of length one specifying the size of the sample from the fitted joint distributions used to estimate the density along an isoline. Samples are collected from the two joint distribution with proportions consistent with the total number of exteme events conditioned on each variable.}

\item{N_Ensemble}{Numeric vector of length one specifying the number of possible design events sampled along the isoline of interest.}

\item{Con2}{Character vector of length one specifying the name of variable in the second column of \code{Data}.}
}
\value{
Plot of all the observations (grey circles) as well as the declustered excesses above Thres1 (blue circles) or Thres2 (blue circles), observations may belong to both conditional samples. Also shown is the isoline associated with \code{RP} contoured according to their relative probability of occurrence on the basis of the sample from the two joint distributions, the "most likely" design event (black diamond), and design event under the assumption of full dependence (black triangle) are also shown in the plot. The function also returns a list comprising the design events assuming full dependence \code{"FullDependence"}, as well as once the dependence between the variables is accounted for the "Most likley" {"MostLikelyEvent"} as well as an {"Ensemble"} of possible design events.
}
\description{
Calculates the single design event under the assumption of full dependence, or once accounting for dependence between variables the single "most-likely" or an ensemble of possible design events.
}
\examples{
S22.Rainfall<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                             Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                             Con_Variable="Rainfall",Thres=0.97)
S22.OsWL<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                         Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                         Con_Variable="OsWL",Thres=0.97)
S22.Copula.Rainfall<-Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                        Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],Thres =0.97,
                                        y_lim_min=-0.075,y_lim_max=0.25,
                                        Upper=c(2,9),Lower=c(2,10),GAP=0.15)$Copula_Family_Var1
S22.Copula.OsWL<-Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                    Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],Thres =0.97,
                                    y_lim_min=-0.075, y_lim_max =0.25,
                                   Upper=c(2,9),Lower=c(2,10),GAP=0.15)$Copula_Family_Var2
Design_Event_2D(Data=S22.Detrend.df[,-c(1,4)], Data_Con1=S22.Rainfall$Data,
               Data_Con2=S22.OsWL$Data, Thres1=0.97, Thres2=0.97,
               Copula_Family1=S22.Copula.Rainfall, Copula_Family2=S22.Copula.OsWL,
               Marginal_Dist1="Logis", Marginal_Dist2="Twe",RP=100,N=10,N_Ensemble=10)
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Copula_Threshold_2D}} \code{\link{Diag_Non_Con}}  \code{\link{Diag_Non_Con_Trunc}}
}
