% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design_Event_2D_Grid.R
\name{Design_Event_2D_Grid}
\alias{Design_Event_2D_Grid}
\title{Derives a single or ensemble of bivariate design events}
\usage{
Design_Event_2D_Grid(
  Data,
  Data_Con1,
  Data_Con2,
  u1,
  u2,
  Thres1 = NA,
  Thres2 = NA,
  N_Both,
  Copula_Family1,
  Copula_Family2,
  Marginal_Dist1,
  Marginal_Dist2,
  Marginal_Dist1_Par = NA,
  Marginal_Dist2_Par = NA,
  Con1 = "Rainfall",
  Con2 = "OsWL",
  GPD1 = NULL,
  GPD2 = NULL,
  Rate_Con1 = NA,
  Rate_Con2 = NA,
  Tab1 = NULL,
  Tab2 = NULL,
  mu = 365.25,
  GPD_Bayes = FALSE,
  Decimal_Place = 2,
  Grid_x_min = NA,
  Grid_x_max = NA,
  Grid_y_min = NA,
  Grid_y_max = NA,
  Grid_x_interval = NA,
  Grid_y_interval = NA,
  RP,
  x_lab = "Rainfall (mm)",
  y_lab = "O-sWL (mNGVD 29)",
  x_lim_min = NA,
  x_lim_max = NA,
  y_lim_min = NA,
  y_lim_max = NA,
  Isoline_Probs = "Sample",
  N = 10^6,
  N_Ensemble = 0,
  Sim_Max = 10,
  Plot_Quantile_Isoline = FALSE
)
}
\arguments{
\item{Data}{Data frame of dimension \code{nx2} containing two co-occurring time series of length \code{n}.}

\item{Data_Con1}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column.}

\item{Data_Con2}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column. Can be obtained using the \code{Con_Sampling_2D} function.}

\item{u1}{Numeric vector of length one specifying the threshold, expressed as a quantile, above which the variable in the first column was sampled in \code{Data_Con1}.}

\item{u2}{Numeric vector of length one specifying the threshold, expressed as a quantile, above which the variable in the second column was sampled in \code{Data_Con2}.}

\item{Thres1}{Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in \code{Data_Con1}. Only one of \code{u1} and \code{Thres1} should be supplied. Default is \code{NA}.}

\item{Thres2}{Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in \code{Data_Con2}. Only one of \code{u2} and \code{Thres2} should be supplied. Default is \code{NA}.}

\item{N_Both}{Numeric vector of length one specifying the number of data points that feature in both conditional samples.}

\item{Copula_Family1}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con1} dataset.}

\item{Copula_Family2}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con2} dataset. Best fitting of 40 copulas can be found using the \code{Copula_Threshold_2D} function.}

\item{Marginal_Dist1}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con1}.}

\item{Marginal_Dist2}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con2}.}

\item{Marginal_Dist1_Par}{Object containing the distribution fitted to the non-conditioned variable in \code{Data_Con1}. For example it could be of class \code{fitdistr}. Default is \code{NA} as distributions specified by \code{Marginal_Dist1} are fitted within the function.}

\item{Marginal_Dist2_Par}{Object containing the distribution fitted to the non-conditioned variable in \code{Data_Con2}. For example it could be of class \code{fitdistr}. Default is \code{NA} as distributions specified by \code{Marginal_Dist2} are fitted within the function.}

\item{Con1}{Character vector of length one specifying the name of variable in the first column of \code{Data}.}

\item{Con2}{Character vector of length one specifying the name of variable in the second column of \code{Data}.}

\item{GPD1}{Output of \code{GPD_Fit} applied to variable \code{con1} i.e., GPD fit \code{con1}. Default \code{NULL}. Only one of \code{u1}, \code{Thres1}, \code{GPD1} and \code{Tab1} is required.}

\item{GPD2}{Output of \code{GPD_Fit} applied to variable \code{con2} i.e., GPD fit \code{con2}. Default \code{NULL}. Only one of \code{u2}, \code{Thres2}, \code{GPD2} and \code{Tab2} is required.}

\item{Rate_Con1}{Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con1}. Default is \code{NA}.}

\item{Rate_Con2}{Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con2}. Default is \code{NA}.}

\item{Tab1}{Data frame specifying the return periods of variable \code{con1}, when conditioning on \code{con1}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con1} for the inter-arrival time (1/rate) of the sample. Default is \code{NULL}.}

\item{Tab2}{Data frame specifying the return periods of variable \code{con2}, when conditioning on \code{con2}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con2} for the inter-arrival time (1/rate) of the sample. Default is \code{NULL}.}

\item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in \code{Data}. Default is \code{365.25}, daily data.}

\item{GPD_Bayes}{Logical; indicating whether to use a Bayesian approach to estimate GPD parameters. This involves applying a penalty to the likelihood to aid in the stability of the optimization procedure. Default is \code{FALSE}.}

\item{Decimal_Place}{Numeric vector specifying the number of decimal places to which to specify the isoline. Default is \code{2}}

\item{Grid_x_min}{Numeric vector of length one specifying the minimum value of the variable in first column of \code{Data} contained in the grid.}

\item{Grid_x_max}{Numeric vector of length one specifying the maximum value of the variable in first column of \code{Data} contained in the grid.}

\item{Grid_y_min}{Numeric vector of length one specifying the minimum value of the variable in second column of \code{Data} contained in the grid.}

\item{Grid_y_max}{Numeric vector of length one specifying the maximum value of the variable in second column of \code{Data} contained in the grid.}

\item{Grid_x_interval}{Numeric vector of length one specifying the resolution of the grid in terms of the variable in first column of \code{Date}. Default is an interval \code{2} of between consecutive values.}

\item{Grid_y_interval}{Numeric vector of length one specifying the resolution of the grid in terms of the variable in second column of \code{Date}. Default is an interval \code{0.1} of between consecutive values.}

\item{RP}{Numeric vector specifying the return periods of interest.}

\item{x_lab}{Character vector specifying the x-axis label.}

\item{y_lab}{Character vector specifying the y-axis label.}

\item{x_lim_min}{Numeric vector of length one specifying x-axis minimum. Default is \code{NA}.}

\item{x_lim_max}{Numeric vector of length one specifying x-axis maximum. Default is \code{NA}.}

\item{y_lim_min}{Numeric vector of length one specifying y-axis minimum. Default is \code{NA}.}

\item{y_lim_max}{Numeric vector of length one specifying y-axis maximum. Default is \code{NA}.}

\item{Isoline_Probs}{Character vector of length one specifying whether to calculate relative probabilities of points on the isoline from a \code{"Sample"} simulated from the fitted copula models or from the \code{"Observations"}.Default is \code{"Sample"}.}

\item{N}{Numeric vector of length one specifying the size of the sample from the fitted joint distributions used to estimate the density along an isoline. Samples are collected from the two joint distribution with proportions consistent with the total number of extreme events conditioned on each variable. Default is \code{10^6}}

\item{N_Ensemble}{Numeric vector of length one specifying the number of possible design events sampled along the isoline of interest.}

\item{Sim_Max}{Numeric vector of length one specifying the maximum value, given as a multiple of the largest observation of each variable, permitted in the sample used to estimate the (relative) probabilities along the isoline.}

\item{Plot_Quantile_Isoline}{Logical; indicating whether to first plot the quantile isoline. Default is \code{FALSE}.}
}
\value{
Plot of all the observations (grey circles) as well as the declustered excesses above Thres1 (blue circles) or Thres2 (blue circles), observations may belong to both conditional samples. Also shown is the isoline associated with \code{RP} contoured according to their relative probability of occurrence on the basis of the sample from the two joint distributions, the "most likely" design event (black diamond), and design event under the assumption of full dependence (black triangle) are also shown in the plot. The function also returns a list comprising the design events assuming full dependence \code{"FullDependence"}, as well as once the dependence between the variables is accounted for the "Most likley" \code{"MostLikelyEvent"} as well as an \code{"Ensemble"} of possible design events and relative probabilities of events on the isoline \code{Contour}. The quantile isolines with \code{Quantile_Isoline_1} and \code{Quantile_Isoline_2}, and GPD thresholds with \code{Threshold_1} and \code{Threshold_2}.
}
\description{
Calculates the isoline and relative probability of events on the isoline, given the observational data, for one or more user-specified return periods. Outputs the single "most-likely" design event or an ensemble of possible design events obtained by sampling along the isoline according to these relative probabilities. The design event under the assumption of full dependence is also computed. Isoline is derived by calculating annual exceedance probabilities from both copula models on a user-specified grid rather by overlaying the partial isolines from the two copula models as in \code{Design_Event_2D}.
}
\examples{
S22.Rainfall<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                             Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                             Con_Variable="Rainfall",u=0.97)
S22.OsWL<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                         Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
                         Con_Variable="OsWL",u=0.97)
S22.Copula.Rainfall<-Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                        Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],u1 =0.97,
                                        y_lim_min=-0.075,y_lim_max=0.25,
                                        Upper=c(2,9),Lower=c(2,10),GAP=0.15)$Copula_Family_Var1
S22.Copula.OsWL<-Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
                                    Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],u2 =0.97,
                                    y_lim_min=-0.075, y_lim_max =0.25,
                                    Upper=c(2,9),Lower=c(2,10),GAP=0.15)$Copula_Family_Var2
Design.Event<-Design_Event_2D_Grid(Data=S22.Detrend.df[,-c(1,4)],
                                  Data_Con1=S22.Rainfall$Data, Data_Con2=S22.OsWL$Data,
                                  u1=0.97, u2=0.97, N_Both=3,
                                  Copula_Family1=S22.Copula.Rainfall,
                                  Copula_Family2=S22.Copula.OsWL,
                                  Marginal_Dist1="Logis", Marginal_Dist2="Twe",
                                  RP=c(5,100),
                                  Grid_x_interval=0.1,Grid_y_interval=0.1,
                                  N=10^6,N_Ensemble=10,
                                  Plot_Quantile_Isoline=FALSE)
#Extracting the 100-year isoline from the output
Design.Event$`100`$Isoline
}
\seealso{
\code{\link{Copula_Threshold_2D}} \code{\link{Diag_Non_Con}} \code{\link{Diag_Non_Con_Trunc}}
}
