% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detrend.R
\name{Detrend}
\alias{Detrend}
\title{Detrends a time series.}
\usage{
Detrend(
  Data,
  Method = "window",
  Window_Width = 89,
  End_Length = 1826,
  PLOT = FALSE,
  x_lab = "Date",
  y_lab = "Data"
)
}
\arguments{
\item{Data}{Data frame containing two columns. In column: \itemize{
\item \code{1} A \code{"Date"} object of equally spaced discrete time steps.
\item \code{2} Numeric vector containing corresponding time series values. No \code{NA}s allowed.
}}

\item{Method}{Character vector of length one specifying approach used to detrend the data. Options are moving average \code{"window"} (default) and \code{"linear"}.}

\item{Window_Width}{Numeric vector of length one specifying length of the moving average window. Default is \code{89}, window comprises the observation plus 44 days either side, which for daily data corresponds to an approximate 3 month window.}

\item{End_Length}{Numeric vector of length one specifying number of observations at the end of the time series used to calculate the present day average. Default is \code{1826}, which for daily data corresponds to the final five years of observations.}

\item{PLOT}{Logical; whether to plot original and detrended series. Default is \code{"FALSE"}.}

\item{x_lab}{Character vector of length one specifying x-axis label. Default is \code{"Date"}.}

\item{y_lab}{Character vector of length one specifying y-axis label. Default is \code{"Data"}.}
}
\value{
Numeric vector of the detrended time series.
}
\description{
Detrends a time series using either a linear fit covering the entire dataset or moving average trend correction with a user-specified window width.
}
\examples{
\donttest{
#Detrending ocean-side water level at site S22 using a 3 month moving average window and the last
#five years of observations to calculate the present day average.
data("S22_T_MAX_Daily_Completed")
Detrend(S22_T_MAX_Daily_Completed[,1:2],Method = "window",Window_Width= 89,
        End_Length = 1826, PLOT=FALSE,x_lab="Data",y_lab="Data")
}
}
