% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag_Non_Con_Sel.R
\name{Diag_Non_Con_Sel}
\alias{Diag_Non_Con_Sel}
\title{Demonstrate the goodness of fit of the selected non-extreme marginal distribution}
\usage{
Diag_Non_Con_Sel(
  Data,
  Omit = NA,
  x_lab = "Data",
  y_lim_min = 0,
  y_lim_max = 1,
  Selected
)
}
\arguments{
\item{Data}{Numeric vector containing realizations of the variable of interest.}

\item{Omit}{Character vector specifying any distributions that are not to be tested. Default \code{"NA"}, all distributions are fit.}

\item{x_lab}{Numeric vector of length one specifying. Label on the x-axis of histogram and cumulative distribution plot.}

\item{y_lim_min}{Numeric vector of length one specifying the lower y-axis limit of the histogram.}

\item{y_lim_max}{Numeric vector of length one specifying the upper y-axis limit of the histogram.}

\item{Selected}{Character vector of length one specifying the chosen distribution. Options are: Gaussian \code{"Gaus"}, Gumbel \code{"Gum"}, Laplace \code{"Lapl"}, logistic \code{"Logis"}, and reverse Gumbel \code{"RGum"}.}
}
\value{
Panel consisting of three plots. Upper plot: Plots depicting the AIC of the two fitted distributions. Middle plot: Probability Density Functions (PDFs) of the \code{selected} distributions superimposed on a histogram of the data. Lower plot: Cumulative distribution function (CDFs) of the \code{selected} distribution overlaid on a plot of the empirical CDF.
}
\description{
Plots demonstrating the goodness of fit of a selected (not truncated) non-extreme marginal distribution to a dataset.
}
\examples{
S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
                              Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
                              Con_Variable="Rainfall",Thres=0.97)
Diag_Non_Con(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
             y_lim_min=0,y_lim_max=1.5)
Diag_Non_Con_Sel(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
                 y_lim_min=0,y_lim_max=1.5,Selected="Logis")
}
\seealso{
\code{\link{Diag_Non_Con}}
}
