% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{G_3355}
\alias{G_3355}
\title{Groundwater Well G-3355}
\format{
A data frame with 12,340 rows and 1 variable:
\describe{
\item{Date}{Date of observation (YYYY-MM-DD)}
\item{Value}{Groudwater elevation}
}
}
\source{
South Florida Water Management District, via \url{https://www.sfwmd.gov/science-data/dbhydro/}
}
\usage{
data(G_3355)
}
\description{
Time series of daily water elevation at Well G-3355.
}
\keyword{datasets}
