% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{G_580A}
\alias{G_580A}
\title{Groundwater Levels at Well G-580A}
\format{
A data frame with 12,301 rows and 2 variables:
\describe{
\item{Date}{Date of observation (YYYY-MM-DD)}
\item{Value}{Groundwater elevation (ft NGVD29)}
}
}
\source{
South Florida Water Management District, via
\url{https://www.sfwmd.gov/science-data/dbhydro/}
}
\usage{
data(G_580A)
}
\description{
Time series of daily water elevation at Well G-580A.
}
\keyword{datasets}
