% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD_Fit.R
\name{GPD_Fit}
\alias{GPD_Fit}
\title{Fits a single generalised Pareto distribution - Fit}
\usage{
GPD_Fit(Data, Data_Full, u = 0.95, PLOT = FALSE, xlab_hist = "Data",
  y_lab = "Data")
}
\arguments{
\item{Data}{Numeric vector containing the declusted data.}

\item{Data_Full}{Numeric vector containing the non-declustered data.}

\item{u}{GPD threshold; as a quantile \code{[0,1]} of \code{Data} vector. Default is \code{0.95}.}

\item{xlab_hist}{Character vector of length one. Histogram x-axis label. Default is \code{"Data"}.}

\item{y_lab}{Character vector of length one. Histogram x-axis label. Default is \code{"Data"}.}

\item{Plot}{Logical; indicating whether to plot diagnostics. Default is \code{FALSE}.}
}
\value{
List comprising the GPD \code{Threshold}, shape parameter \code{xi} and scale parameters \code{sigma} along with their standard errors \code{sigma.SE} and \code{xi.SE}.
}
\description{
Fit a Generalized Pareto Distribution (GPD) to a declstered dataset.
}
\section{Details}{

The fitted GPD model, is following parameterised as follows: \eqn{P( X > x| X > u)}
}

\examples{
Decluster(Data=S28_T_MAX_Daily_Completed_Detrend$Detrend)
}
