% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD_Parameter_Stability_Plot.R
\name{GPD_Parameter_Stability_Plot}
\alias{GPD_Parameter_Stability_Plot}
\title{GPD parameter stability plots}
\usage{
GPD_Parameter_Stability_Plot(Data, Data_Full, u = 0.95, PLOT = FALSE,
  xlab_hist = "Data", y_lab = "Data")
}
\arguments{
\item{Data}{Numeric vector containing the declusted data.}

\item{Data_Full}{Numeric vector containing the non-declustered data.}

\item{u}{Numeric vector of GPD thresholds; given as a quantiles \code{[0,1]} of \code{Data} vector. Default is \code{0.9} to \code{0.999} in intervals of \code{0.001}.}

\item{Plot}{Logical; indicating whether to plot diagnostics. Default is \code{FALSE}.}
}
\value{
Plot of the shape and modified scale parameter estimates along with their errors bars over the range of specified thresholds.
}
\description{
Plots showing the stability of the GPD scale and shape parameter estimates across a specified range of thresholds.
}
\examples{
GPD_Parameter_Stability_Plot(Data = S20.Detrend.Declustered.df$Rainfall,
                             Data_Full= na.omit(S20.Detrend.df$Rainfall),
                             u=seq(0.9,0.999,0.001))
}
\seealso{
\code{\link{Decluster}}
}
