% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT04.R
\name{HT04}
\alias{HT04}
\title{Fits and simulates from the conditional multivariate approach of Heffernan and Tawn (2004)}
\usage{
HT04(data_Detrend_Dependence_df, data_Detrend_Declustered_df, u_Dependence,
  Migpd, mu = 365.25, N = 100, Margins = "gumbel", V = 10,
  Maxit = 10000)
}
\arguments{
\item{data_Detrend_Dependence_df}{A dataframe with (n+1) columns, containing in column \itemize{
\item 1 - Continuous sequence of dates spanning the first to the final time of any of the variables are recorded.
\item 2:(n+1) - Values, detrended where necessary, of the variables to be modelled.
}}

\item{data_Detrend_Declustered_df}{A dataframe with (n+1) columns, containing in column \itemize{
\item 1 - Continuous sequence of dates spanning the first to the final time of any of the variables are recorded.
\item 2:(n+1) - Declustered and if necessary detrended values of the variables to be modelled.
}}

\item{u_Dependence}{Dependence quantile. Specifies the (sub-sample of) data to which the dependence model is fitted, that for which the conditioning variable exceeds the threshold associated with the presecribed quantile. Default is \code{0.7}, thus the dependence parameters are estimated using the data with the highest \code{30\%} of values of the conditioning variables.}

\item{Migpd}{An \code{Migpd} object, containing the generalised Pareto models fitted (independently) to each of the variables.}

\item{Margins}{Character vector specifying the form of margins to which the data are transformed for carrying out dependence estimation. Default is \code{"gumbel"}, alternative is \code{"laplace"}. Under Gumbel margins, the estimated parameters \code{a} and \code{b} describe only positive dependence, while \code{c} and \code{d} describe negative dependence in this case. For Laplace margins, only parameters \code{a} and \code{b} are estimated as these capture both positive and negative dependence.}

\item{V}{See documentation for mexDependence.}

\item{Maxit}{See documentation for mexDependence.}
}
\value{
List comprising the fitted HT04 models \code{Models}, proportion of the time each variable is most extreme, given at least one variable is extreme \code{Prop}, as well as the simulated values on the transformed \code{u.sim} and original {x.sim} scales.
}
\description{
Fitting and simulating the conditional multivariate approach of Heffernan and Tawn (2004) to a dataset comprising 3 variables. Function utilizes the \code{mexDependence} and \code{} functions from the \code{texmex} package.
}
\examples{
HT04(data_Detrend_Dependence_df = S22.Detrend.df,data_Detrend_Declustered_df = S22.Detrend.Declustered.df ,Migpd = S22_GPD, u_Dependence=0.7,Margins = "gumbel")
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Migpd_Fit}}
}
