% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HURDAT.R
\name{HURDAT}
\alias{HURDAT}
\title{HURDAT}
\usage{
HURDAT(Data, lat.loc, lon.loc, rad)
}
\arguments{
\item{Data}{A data frame where the first column is a date or date_time object.}

\item{lat.loc}{Numeric vector of length one specifying the latitude (in degrees)of the location of interest.}

\item{lon.loc}{Numeric vector of length one specifying the longitude (in degrees) of the location of interest.}

\item{rad}{Vector of length one, specifying radius (in km) about the location of interest to report storm names.}
}
\value{
The data frame \code{Data} with an additional column containing the named storms.
}
\description{
Appends a data frame with the names of any storms in the HURDAT database with a center of circulation within a specified radius of a location.
}
