% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HURDAT2}
\alias{HURDAT2}
\title{Atlantic hurricane database (HURricane DATa 2nd generation) - HURDAT2)}
\format{
A data frame with columns parsed from the fixed-width HURDAT2 format:
\describe{
\item{V1}{See \url{https://www.nhc.noaa.gov/data/hurdat/hurdat2-format-atl-1851-2021.pdf} for description.}
}
}
\source{
National Hurricane Center, via
\url{https://www.nhc.noaa.gov/data/}
}
\usage{
data(HURDAT2)
}
\description{
"Best-track" data at a six-hourly resolution for all known tropical cyclones and subtropical cyclones available from the b-decks in the Automated Tropical Cyclone Forecast (ATCF – Sampson and Schrader 2000) system database.
}
\keyword{datasets}
