% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intensity.R
\name{Intensity}
\alias{Intensity}
\title{Intensity}
\usage{
Intensity(Data, Cluster_Max, Base_Line = "Mean")
}
\arguments{
\item{Data}{A data frame containing the water level time series. First column may be a \code{"Date"} object.}

\item{Cluster_Max}{Numeric vector containing indexes of the water level peaks in \code{Data}. If analyzing a sample conditioned on water level derived using \code{Con_Sample_2D()} set equal to the \code{$xcon} output.}

\item{Base_Line}{Vector of length one, specifying water level about which to calculate the intensity. Default is \code{"Mean"} where the mean of the entire time series is used as the baseline water level above which intensity is calculated.}
}
\value{
A data frame with the following columns:  \itemize{
\item \code{Pre.High} Index of the water level column of \code{Data} containing the preceding high water level.
\item \code{Fol.High} Index of the water level column of \code{Data} containing the following high water level.
\item \code{Pre.Low} Index of the water level column of \code{Data} containing the preceding  low water level.
\item \code{Fol.Low} Index of the water level column of \code{Data} containing the following low water level.
\item \code{Intensity} Intensity of the extreme water level event.
}
}
\description{
Calculates the "intensity" of extreme water levels, as defined in Wahl et al. (2011).
}
\examples{
#Decluster O-sWL series at S-13 using a runs method
S13.OsWL.Declust = Decluster(Data=S13.Detrend.df$OsWL,
                            SepCrit=24*7, u=0.99667)
#Calculate O-sWL of the identified cluster maximum
intensity = Intensity(Data=S13.Detrend.df[,c(1,3)],Cluster_Max=S13.OsWL.Declust$EventsMax)
#Plot O-sWL series identifying cluster maximum (in red) and print "intensity" above each maximum
plot(as.Date(S13.Detrend.df$Date_Time),
     S13.Detrend.df$OsWL)
points(as.Date(S13.Detrend.df$Date_Time[S13.OsWL.Declust$EventsMax]),
       S13.Detrend.df$OsWL[S13.OsWL.Declust$EventsMax],pch=16,col=2)
text(as.Date(S13.Detrend.df$Date_Time[S13.OsWL.Declust$EventsMax]),
     S13.Detrend.df$OsWL[S13.OsWL.Declust$EventsMax]+0.2,
     round(intensity$Intensity,0),cex=0.5)
}
\seealso{
\code{\link{Decluster}} \code{\link{WL_Curve}}
}
