% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Migpd_Fit.R
\name{Migpd_Fit}
\alias{Migpd_Fit}
\title{Fits Multiple independent generalized Pareto models - Fit}
\usage{
Migpd_Fit(Data, mth, mqu, penalty = "gaussian", maxit = 10000,
  trace = 0, verbose = FALSE, priorParameters = NULL)
}
\arguments{
\item{Data}{A dataframe with \code{n} columns, each comprising a declustered and if necessary detrended time series to be modelled.}

\item{mth}{Marginal thresholds, above which generalized Pareto models are fitted. Numeric vector of length \code{n}.}

\item{mqu}{Marginal quantiles, above which generalized Pareto models are fitted. \strong{Only one of \code{mth} and \code{mqu} should be supplied.} Numeric vector of length \code{n}.}

\item{penalty}{See \code{\link{ggplot.migpd}}.}

\item{maxit}{See \code{\link{ggplot.migpd}}.}

\item{trace}{See \code{\link{ggplot.migpd}}.}

\item{verbose}{See \code{\link{ggplot.migpd}}.}

\item{priorParameters}{See \code{\link{ggplot.migpd}}.}
}
\value{
An object of class \code{"migpd"}. There are \code{coef}, \code{print}, \code{plot}, \code{ggplot} and \code{summary} functions available.
}
\description{
Fit multiple independent generalized Pareto models to each column of a dataframe. Edited version of the \code{migpd} function in \code{texmex}, to alllow for \code{NA}s in a time series.
}
\examples{
#With date as first column
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df, mqu =c(0.99,0.99,0.99))
#Without date as first column
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1], mqu =c(0.99,0.99,0.99))
}
\seealso{
\code{\link{Decluster}} \code{\link{Detrend}} \code{\link{Dataframe_Combine}}
}
