% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_SLR.R
\name{NOAA_SLR}
\alias{NOAA_SLR}
\title{NOAA sea-level rise scenarios}
\usage{
NOAA_SLR(
  OsWL_req,
  SLR_scen = c("High", "Intermediate", "Low"),
  Input_unit = "m",
  Year.Inital = 2020
)
}
\arguments{
\item{OsWL_req}{Numeric vector of SLR required.}

\item{SLR_scen}{Character vector specifying which of the NOAA (2017) scenarios to consider. Options include \code{High}, Intermediate high \code{Int.High}, \code{Intermediate}, Intermediate low (\code{Int.Low}) and \code{Low}.}

\item{Input_unit}{Character vector of length one; specifying units of SLR. Default is meters \code{"m"}, other option is feet \code{"ft"}.}

\item{Year}{Character vector of length one; specifying}
}
\value{
List comprising the specified \code{Threshold} as the quantile of the conditioning variable above which declustered excesses are paired with co-occurrences of the other variable, the resulting two-dimensional sample \code{data} and \code{name} of the conditioning variable.
}
\description{
Time (in years) for a specified amount of sea-level rise (SLR) to occur at Miami Beach according to the five SLR scenarios in NOAA 2017 report titled "Global and Regional Sea Level Rise Scenarios for the United States".
}
\examples{
NOAA_SLR<-function(OsWL_req=seq(0,1,0.01),SLR_scen = c("High","Intermediate","Low"),
                   Input_unit="m")
}
