% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOAAetal2012}
\alias{NOAAetal2012}
\title{2012 NOAA Sea Level Rise Projections for Virginia Key}
\format{
A data frame with XX rows and 1 variable:
\describe{
\item{Date}{Date (YYYY-MM-DD)}
\item{Value}{Projected sea level rise (inches)}
}
}
\source{
NOAA National Ocean Service. Report available at:
\url{https://repository.library.noaa.gov/view/noaa/11124}
}
\usage{
data(NOAAetal2012)
}
\description{
Sea level rise projections for Virginia Key under four emission scenarios from NOAA Technical Report OAR CPO-1 (2012).
}
\keyword{datasets}
