% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOAAetal2017}
\alias{NOAAetal2017}
\title{2017 NOAA Sea Level Rise Projections for Virginia Key}
\format{
A data frame with XX rows and 1 variable:
\describe{
\item{Date}{Date (YYYY-MM-DD)}
\item{Value}{Projected sea level rise (inches)}
}
}
\source{
NOAA National Ocean Service. Report available at: \url{https://tidesandcurrents.noaa.gov/publications/techrpt83_Global_and_Regional_SLR_Scenarios_for_the_US_final.pdf}
}
\usage{
data(NOAAetal2017)
}
\description{
Sea level rise projections for Virginia Key derived from NOAA Technical Report NOS CO-OPS 083 (2017). The report defines six global mean sea level rise scenarios, which are regionally downscaled on a 1-degree grid covering U.S. coastlines.
}
\keyword{datasets}
