% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOAAetal2022}
\alias{NOAAetal2022}
\title{2022 NOAA Global and Regional Sea Level Rise Scenarios for the United States}
\format{
A data frame with XX rows and 1 variable:
\describe{
\item{Date}{Date of observation (YYYY-MM-DD)}
\item{Value}{Rainfall totals (Inches)}
}
}
\source{
NOAA National Ocean Service, via \url{https://oceanservice.noaa.gov/hazards/sealevelrise}
}
\usage{
data(NOAAetal2022)
}
\description{
Sea level rise scenarios for Virginia Key from given in NOAA Technical Report NOS 01. In the report, five global mean sea level rise scenarios are used to derive probabilistic regional RSL responses on a 1-degree grid covering the coastlines of the U.S. mainland.
}
\keyword{datasets}
