% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsWL_Intensity.R
\name{OsWL_Intensity}
\alias{OsWL_Intensity}
\title{Ocean-side Water Level Intensity}
\usage{
OsWL_Intensity(
  Data,
  Cluster_Max,
  Base_Line = mean(Data$OsWL, na.rm = T),
  Rainfall_Interval = 24
)
}
\arguments{
\item{Data}{A data frame with co-occurring rainfall and O-sWL time series in two columns labeled \code{"Rainfall"} and \code{"OsWL"}, respectively.}

\item{Cluster_Max}{Numeric vector containing indexes of peaks in the O-sWL column of \code{Data}. If analyzing a sample conditioned on O-sWL derived using \code{Con_Sample_2D()} set equal to the \code{$xcon} output.}

\item{Base_Line}{Numeric vector of length one, specifying water level about which to calculate the intensity. Default is the mean O-sWL.}

\item{Rainfall_Interval}{Numeric vector of length one, specifying length of time before and after a peak over which to sum rainfall totals. Total window width is 2*\code{Rainfall_Interval}+1. Default is \code{24}.}
}
\value{
A data frame with the following columns:  \itemize{
\item \code{Pre.High} Index of the OsWL column of \code{Data} containing the preceding high water level.
\item \code{Fol.High} Index of the OsWL column of \code{Data} containing the following high water level.
\item \code{Pre.Low} Index of the OsWL column of \code{Data} containing the preceding  low water level.
\item \code{Fol.Low} Index of the OsWL column of \code{Data} containing the following low water level.
\item \code{Intensity} Intensity of the O-sWL.
\item \code{V} Total rainfall volume within \code{Rainfall_Interval} before and after the peak.
}
}
\description{
Calculates the "intensity" of extreme water levels, as defined in Wahl et al. (2011).
}
\examples{
#Decluster O-sWL series at S-13 using a runs method
S13.OsWL.Declust = Decluster(Data=S13.Detrend.df$OsWL,
                            SepCrit=24*7, u=0.99667)
#Calculate O-sWL of the identified cluster maximum
intensity = OsWL_Intensity(Data=S13.Detrend.df,Cluster_Max=S13.OsWL.Declust$EventsMax)
#Plot O-sWL series identifying cluster maximum (in red) and print "intensity" above each maximum
plot(S13.Detrend.df$Date_Time,
     S13.Detrend.df$OsWL)
points(S13.Detrend.df$Date_Time[S13.OsWL.Declust$EventsMax],
       S13.Detrend.df$OsWL[S13.OsWL.Declust$EventsMax],pch=16,col=2)
text(S13.Detrend.df$Date_Time[S13.OsWL.Declust$EventsMax],
     S13.Detrend.df$OsWL[S13.OsWL.Declust$EventsMax]+0.2,
     round(intensity$Intensity,0),cex=0.5)
}
\seealso{
\code{\link{Decluster}} \code{\link{WL_Curve}}
}
