% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Perrine_df}
\alias{Perrine_df}
\title{Rainfall Totals at Perrine, FL.}
\format{
A data frame with XX rows and 2 variables:
\describe{
\item{Date}{Date of observation (YYYY-MM-DD)}
\item{Value}{Rainfall totals (inches)}
}
}
\source{
NOAA National Centers for Environmental Information, via
\url{https://www.ncdc.noaa.gov/}
}
\usage{
data(Perrine_df)
}
\description{
Time series of daily rainfall totals from the gauge at Perrine 4W, FL (Network:ID GHCND:USC00087020).
}
\keyword{datasets}
