% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PVAL_AU2_LMOM_4}
\alias{PVAL_AU2_LMOM_4}
\title{Automatic Peaks Over Threshold (POT) Threshold Selection}
\format{
A data frame or matrix of p-values, where rows correspond to \link{describe what they represent} and columns to \link{describe columns}.
}
\source{
Supplied by the authors of Solari et al. (2017): \doi{10.1002/2016WR019426}
}
\usage{
PVAL_AU2_LMOM_4
}
\description{
P-values used in the Automatic Peaks Over Threshold (POT) threshold selection procedure in Solari et al. (2017) for a subset of shape and scale parameter values.
}
\references{
Solari, et al. (2017). \url{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2016WR019426}
}
\keyword{datasets}
