% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{S13_Rainfall}
\alias{S13_Rainfall}
\title{Rainfall Totals at S-13}
\format{
A data frame with XX rows and 2 variables:
\describe{
\item{Date_Time}{Time of observation (YYYY-MM-DD HH:MM:SS)}
\item{Rainfall}{Rainfall totals (inches)}
}
}
\source{
South Florida Water Management District, via
\url{https://www.sfwmd.gov/science-data/dbhydro/}
}
\usage{
data(S13_Rainfall)
}
\description{
Hourly time series of rainfall totals from the gauge at control structure S-13.
}
\keyword{datasets}
