% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{S20.Detrend.Declustered.df}
\alias{S20.Detrend.Declustered.df}
\title{Declustered Time Series for Case Study Site S-20 in Jane et al. (2020)}
\format{
A data frame with XX rows and 4 variables:
\describe{
\item{Date}{Date of observation (YYYY-MM-DD)}
\item{Rainfall}{Rainfall totals (inches)}
\item{OsWL}{Ocean-side water levels (ft NGVD29)}
\item{Groundwater}{Groundwater levels (ft NGVD29)}
}
}
\source{
Rainfall totals from the NOAA National Centers for Environmental Information: \url{https://www.ncdc.noaa.gov/}
Ocean-side water levels and groundwater levels from the South Florida Water Management District: \url{https://www.sfwmd.gov/science-data/dbhydro/}
}
\usage{
data(S20.Detrend.Declustered.df)
}
\description{
Time series of rainfall totals from the Perrine 4W gauge, ocean-side water levels at control structure S-20, and groundwater at Well G-3356 (with missing values interpolated using groundwater levels at Well G-3355).
All time series were declustered using a peaks-over-threshold approach (runs method with a 0.98 quantile threshold and a 3-day separation criterion).
Ocean-side water levels and groundwater levels were each detrended using a 3-month moving average window prior to declustering.
}
\references{
Jane, R., Cadavid, L., Obeysekera, J., and Wahl, T.: Multivariate statistical modelling of the drivers of compound flood events in south Florida, Nat. Hazards Earth Syst. Sci., 20, 2681–2699. \url{https://doi.org/10.5194/nhess-20-2681-2020}
}
\keyword{datasets}
