% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_Scenarios.R
\name{SLR_Scenarios}
\alias{SLR_Scenarios}
\title{Sea level rise scenarios in the Southeast Florida Regional Climate Change Compact:}
\usage{
SLR_Scenarios(SeaLevelRise, Unit = "m")
}
\arguments{
\item{SeaLevelRise}{Numeric vector of length one, sea level rise required.}

\item{data}{A data frame with \code{n} columns, each comprising a declustered and if necessary detrended time series to be modelled.}
}
\value{
An object of class \code{"migpd"}. There are \code{coef}, \code{print}, \code{plot}, \code{ggplot} and \code{summary} functions available.
}
\description{
Calculates and plots time required for sea level rise to reach a specified level according to the three scenarios in the Compact.
}
\examples{
SLRScenarios(0.45)
}
