% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_Copula_Fit.R
\name{Standard_Copula_Fit}
\alias{Standard_Copula_Fit}
\title{Fit an Archimedean/elliptic copula model - Fit}
\usage{
Standard_Copula_Fit(Data, Copula_Type = "Gaussian")
}
\arguments{
\item{Data}{Dataframe containing \code{n} at least partially concurrent time series. First column may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}

\item{Copula_Type}{Type of elliptical copula to be fitted, options are \code{"Gaussian"} (Default), \code{"tcopula"}, \code{"Gumbel"}, \code{"Clayton"} and \code{"Frank"}.}
}
\value{
List comprising the \code{Copula_Type} and the fitted copula \code{Model} object.
}
\description{
Fit a n-dimensional Archimedean or elliptic copula model. Function is simply a repackaging of the \code{fitCopula} function in the \code{copula} package.
}
\examples{
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Gaussian")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="tcopula")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Gumbel")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Clayton")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Frank")
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Standard_Copula_Sel}} \code{\link{CDVineCopSelect}}  \code{\link{BiCopSelect}}
}
