% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_Copula_Sel.R
\name{Standard_Copula_Sel}
\alias{Standard_Copula_Sel}
\title{Selecting best fitting standard (elliptical and Archimeadean) copula}
\usage{
Standard_Copula_Sel(Data)
}
\arguments{
\item{Data}{Data frame containing n at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned \code{NA}. First object may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}
}
\value{
Data frame containing copula name in column 1 and associated AIC in column 2.
Parameters are estimated using the \code{fitCopula()} function in \code{copula} package using maximum pseudo-likelihood estimator \code{"mpl"}. See \code{\link{fitCopula}} for a more thorough explanation.
}
\description{
Fits five n-dimensional standard copula to a dataset and returns their corresponding AIC values.
}
\examples{
Standard_Copula_Sel(Data_Detrend=S22.Detrend.df)
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Standard_Copula_Fit}}
}
