% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Plot.R
\name{Time_Series_Plot}
\alias{Time_Series_Plot}
\title{Rainfall and O-sWL time series plots}
\usage{
Time_Series_Plot(
  Rainfall_Series,
  Oswl_Time_Series,
  Sample,
  Con_Variable,
  Buffer = 6,
  Intensity = NA,
  Event_ID = 1:16,
  Row = 4,
  Col = 4,
  Mar = c(4.2, 4.5, 1.5, 3.5)
)
}
\arguments{
\item{Rainfall_Series}{Data frame with rows comprising time series of rainfall totals associated with cluster maximum of the rainfall series.}

\item{Oswl_Time_Series}{Data frame with rows comprising the water level curves associated with the simulated events in \code{Sample}.}

\item{Sample}{Data frame containing the simulated events. Columns (and their names) required by the function are rainfall peak (Rainfall), O-sWL peak (OsWL), their lag time (Lag), and the ID of the sampled rainfall event (samp).}

\item{Con_Variable}{Character vector of length one specifying the conditioning variable of the events in \code{Sample}.}

\item{Buffer}{Numeric vector of length one specifying the extension of the x-axis before and after the rainfall event when \code{Con_Variable == "Rainfall"}. Default is \code{6}.}

\item{Intensity}{Numeric vector specifying the "intensity" of the O-sWL events in \code{Sample}. Default is \code{NA}.}

\item{Event_ID}{Numeric vector specifying the events in \code{Sample} to be plot.}

\item{Row}{Numeric vector of length one specifying the number of rows of subplots in the Figure.}

\item{Col}{Numeric vector of length one specifying the number of columns of subplots in the Figure. Product of \code{Row} and \code{Col} must be equal to or greater than \code{Event_ID}.}

\item{Mar}{Numeric vector of length one specifying the margin at the (bottom,left,top,right) of the subplots. Default is \code{c(4.2,4.5,1.5,3.5)}.}
}
\value{
Figure
}
\description{
Plots a user specified number of synthetic events where at least O-sWL or rainfall peak exceeds a high threshold. .
}
\examples{
Time_Series_Plot(Rainfall_Series=x_n,Oswl_Time_Series=oswl.ts.rainfall,
                 Sample=S13.rainfall.sample,Con_Variable="Rainfall",Event_ID=1:16)
}
\seealso{
\code{\link{C_Sample}} \code{\link{WL_Curve}}
}
