% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{USACE2013}
\alias{USACE2013}
\title{USACE Sea Level Projections for Virginia Key}
\format{
A data frame containing the projections with columns as follows:
\describe{
\item{Year}{Year of Projection (YYYY)}
\item{Low}{Projections for the low emission scenario (m)}
\item{Int}{Projections for the intermediate emission scenario (m)}
\item{High}{Projections for the high emission scenario (m)}
}
}
\source{
U.S. Army Corps. of Engineers, via. USACE Sea Level Change Calculator: \url{http://www.corpsclimate.us/ccaceslcurves.cfm}
}
\usage{
data(USACE2013)
}
\description{
U.S. Army Corps. of Engineers (USACE) 2013 sea level rise projections for Virginia Key, as provided in Engineering Regulation ER 1100-2-8162.
These projections represent future sea level rise scenarios, resulting in global sea level increases of approximately 0.2 meters (Low scenario), 0.5 meters (Intermediate scenario), and 1.5 meters (High scenario) by the year 2100.
}
\keyword{datasets}
