% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vine_Copula_Fit.R
\name{Vine_Copula_Fit}
\alias{Vine_Copula_Fit}
\title{C and D-vine Copula - Fitting}
\usage{
Vine_Copula_Fit(Data)
}
\arguments{
\item{Data}{Data frame containing \code{n} at least partially concurrent time series. First column may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}
}
\value{
List comprising the vine copula \code{Structure}, pair-copula families composing the C- or D-vine copula \code{Family}, its parameters \code{Par} and \code{Par2}.
}
\description{
Fit either a C- or D-vine copula model. Function is a repackaging the \code{RVineStructureSelect} and \code{RVineCopSelect} functions from the \code{VineCopula} package into a single function.
}
\examples{
S20.Vine<-Vine_Copula_Fit(Data=S20.Detrend.df)
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Vine_Copula_Sim}}
}
