% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula_Threshold_2D.R
\name{Copula_Threshold_2D}
\alias{Copula_Threshold_2D}
\title{Copula Selection With threshold 2D - Fit}
\usage{
Copula_Threshold_2D(
  Data_Detrend,
  Data_Declust,
  u1 = seq(0.9, 0.99, 0.01),
  u2 = seq(0.9, 0.99, 0.01),
  PLOT = TRUE,
  x_lim_min = NA,
  x_lim_max = NA,
  y_lim_min = -1,
  y_lim_max = 1,
  Upper = 0,
  Lower = 0,
  GAP = 0.05,
  Legend = TRUE
)
}
\arguments{
\item{Data_Detrend}{Data frame containing two at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned \code{NA}.}

\item{Data_Declust}{Data frame containing two (independently) declustered at least partially concurrent time series. Time steps must be equally spaced, with missing values assigned \code{NA}.}

\item{u1}{A single or sequence of thresholds, given as a quantile of the observations of the variable in the first column of \code{Data_Detrend} when it is used as the conditioning variable. Default, sequence from \code{0.9} to \code{0.99} at intervals of \code{0.01}.}

\item{u2}{A single or sequence of thresholds, given as a quantile of the observations of the variable in the second column of \code{Data_Detrend} when it is used as the conditioning variable. Default, sequence from \code{0.9} to \code{0.99} at intervals of \code{0.01}.}

\item{PLOT}{Logical; whether to plot the results. Default is \code{"TRUE"}.}

\item{x_lim_min}{Numeric vector of length one specifying x-axis minimum. Default is \code{NA}.}

\item{x_lim_max}{Numeric vector of length one specifying x-axis maximum. Default is \code{NA}.}

\item{y_lim_min}{Numeric vector of length one specifying y-axis minimum. Default \code{-1.0}.}

\item{y_lim_max}{Numeric vector of length one specifying y-axis maximum. Default \code{1.0}.}

\item{Upper}{Numeric vector specifying the element number of the \code{u1} argument for which the copula family name label to appear above the corresponding point on the Kendall's tau coefficient vs threshold plot, when conditioning on the variable in column 1. Default is \code{0}.}

\item{Lower}{Numeric vector specifying the element number of the \code{u2} argument for which the copula family name label to appear below the corresponding point on the Kendall's tau coefficient vs threshold plot, when conditioning on the variable in column 2. Default is \code{0}.}

\item{GAP}{Numeric vector of length one specifying the distance above or below the copula family name label appears the corresponding point on the Kendall's tau coefficient vs threshold plot. Default is \code{0.05}.}

\item{Legend}{Logic vector of length one specifying whether a legend should be plotted. Default is \code{TRUE}.}
}
\value{
List comprising: \itemize{
\item \code{Kendalls_Tau1}
Kendall's tau of a sample
\item \code{p_value_Var1}
p-value when testing the null hypothesis \code{H_0: tau=0} i.e. that there is no correlation between the variables
\item \code{N_Var1}
Size of the dataset
\item \code{Copula_Family_Var1}
Best fitting copula for the specified thresholds
}
when the dataset is conditioned on the variable in column 1.
Analogous vectors \code{Kendalls_Tau2},\code{p_value_Var2}, \code{N_Var2} and \code{Copula_Family_Var2} for the specified thresholds when the dataset is conditioned on the variable in column 2.
}
\description{
Declustered excesses of a (conditioning) variable are paired with co-occurences of the other variable before the best fitting bivariate copula is selected, using \code{BiCopSelect} function in the \code{VineCopula} package, for a single or range of thresholds. The procedure is automatically repeated with the variables switched.
}
\examples{
Copula_Threshold_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
                    Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
                    y_lim_min=-0.075, y_lim_max =0.25,
                    Upper=c(6,8), Lower=c(6,8),GAP=0.1)
}
\seealso{
\code{\link{Dataframe_Combine}}
}
