% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataframe_Combine.R
\name{Dataframe_Combine}
\alias{Dataframe_Combine}
\title{Creates a dataframe containing up to five time series}
\usage{
Dataframe_Combine(data.1, data.2, data.3, data.4 = 0, data.5 = 0,
  n = 3, names)
}
\arguments{
\item{n}{Integer \code{1-5} specifying the number of time series. Default is \code{3}.}

\item{data.1:5}{Dataframes with two columns containing in column \itemize{
\item 1 - Continuous sequence of times spaning from the first to the final recorded observations.
\item 2 - Corresponding values detrended where necessary.
}}
}
\value{
A dataframe containing all times from the first to the most up to date reading of any of the variables.
}
\description{
Combines up to five time series, detrended where necessary, into a single dataframe.
}
\examples{
#Reading in data for site S22
Miami_Airport_df<-read.csv("C:\\\\Users\\\\ro327497\\\\Documents\\\\SFWMD\\\\SFWMD Data\\\\S22\\\\Miami_Airport_df.csv")[,-1]
S22_T_MAX_Daily_Completed_Detrend_Declustered<-read.csv("C:\\\\Users\\\\ro327497\\\\Documents\\\\SFWMD\\\\SFWMD Data\\\\S22\\\\S22_T_MAX_Daily_Completed_Detrend_Declustered.csv")[,-1]
G580A_GWValueFilled_Detrend_Declustered<-read.csv("C:\\\\Users\\\\ro327497\\\\Documents\\\\SFWMD\\\\SFWMD Data\\\\S22\\\\G580A_GWValueFilled_Detrend_Declustered.csv")[,-1]
S22.Detrend.Declustered.df<-Dataframe_Combine(data.1<-Miami_Airport_df,data.2<-S22_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,4)],data.3<-G580A_GWValueFilled_Detrend_Declustered[,c(1,4)],names=c("Rainfall","OsWL","Groundwater"))
S22.Detrend.df<-Dataframe_Combine(data.1<-Miami_Airport_df,data.2<-S22_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,3)],data.3<-G580A_GWValueFilled_Detrend_Declustered[,c(1,3)],names=c("Rainfall","OsWL","Groundwater"))
}
\seealso{
\code{\link{Detrend}}
}
