% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag_Non_Con.R
\name{Diag_Non_Con}
\alias{Diag_Non_Con}
\title{Goodness of fit of non-extreme marginal distributions}
\usage{
Diag_Non_Con(Data, x_lab, y_lim_min = 0, y_lim_max = 1)
}
\arguments{
\item{Data}{Numeric vector containing realizations of the variable of interest.}

\item{x_lab}{Character vector of length one specifying the label on the x-axis of histogram and cummulative distribution plot.}

\item{y_lim_min}{Numeric vector of length one specifying the lower y-axis limit of the histogram. Default is \code{0}.}

\item{y_lim_max}{Numericr vector of length one specifying the upper y-axis limit of the histogram. Default is \code{1}.}
}
\value{
Panel consisting of three plots. Upper plot: Plot depicting the AIC of the two fitted distributions. Middle plot: Probabilty Density Functions (PDFs) of the fitted distribtions superimposed on a histgram of the data. Lower plot: Cummulaibre Distribution Functions (CDFs) of the fitted distributions overlaid on a plot of the empirical CDF.
}
\description{
Fits two (unbounded) non-extreme marginal distributions to a dataset and returns three plots demonstrating their relative goodness of fit.
}
\examples{
S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
                              Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
                              Con_Variable="Rainfall",Thres=0.97)
Diag_Non_Con(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
             y_lim_min=0,y_lim_max=1.5)
}
\seealso{
\code{\link{Copula_Threshold_2D}}
}
