% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag_Non_Con_Sel.R
\name{Diag_Non_Con_Sel}
\alias{Diag_Non_Con_Sel}
\title{Demonstrate the goodness of fit of the slected non-extreme marginal distribution}
\usage{
Diag_Non_Con_Sel(Data, x_lab = "Data", y_lim_min = 0, y_lim_max = 1,
  Selected)
}
\arguments{
\item{Data}{Numeric vector containing realizations of the variable of interest.}

\item{x_lab}{Numeric vector of length one specifyingLabel on the x-axis of histogram and cummulative distribution plot.}

\item{y_lim_min}{Numeric vector of length one specifying the lower y-axis limit of the histogram.}

\item{y_lim_max}{Numeric vector of length one specifying the upper y-axis limit of the histogram.}

\item{Selected}{Charactor vector of length one specifying the chosen distribution, options are the Gaussian \code{"Gaus"} and logistic \code{"Logis"}.}
}
\value{
Panel consisting of three plots. Upper plot: Plots depicting the AIC of the two fitted distributions. Middle plot: Probabilty Density Functions (PDFs) of the \code{selected} distribtions superimposed on a histgram of the data. Lower plot: Cummulative distribution function (CDFs) of the \code{selected} distribution overlaid on a plot of the empirical CDF.
}
\description{
Plots demonstrating the goodness of fit of a selected non-extreme marginal distribution to a dataset.
}
\examples{
S22.Rainfall<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)], Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)], Con_Variable="Rainfall",Thres=0.97)
Diag_Non_Con(Data=S22.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD)",y_lim_min=0,y_lim_max=1.5)
Diag_Non_Con_Sel(Data=S22.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD)",y_lim_min=0,y_lim_max=1.5,Selected="Twe")
}
