% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD_Fit.R
\name{GPD_Fit}
\alias{GPD_Fit}
\title{Fits a single generalized Pareto distribution - Fit}
\usage{
GPD_Fit(
  Data,
  Data_Full,
  u = 0.95,
  Thres = NA,
  mu = 365.25,
  GPD_Bayes = TRUE,
  Method = "Standard",
  min.RI = 1,
  max.RI = 100,
  PLOT = FALSE,
  xlab_hist = "Data",
  y_lab = "Data"
)
}
\arguments{
\item{Data}{Numeric vector containing the declusted data.}

\item{Data_Full}{Numeric vector containing the non-declustered data.}

\item{u}{GPD threshold expressed as a quantile \code{[0,1]} of \code{Data} vector. Default is \code{0.95}.}

\item{Thres}{GPD threshold expressed on the original scale of the \code{"Data"}. Only one of \code{u} and \code{Thres} should be supplied. Default is \code{NA}.}

\item{mu}{Numeric vector of length one specifying (average) occurrence frequency of events in the \code{Data_Full} input. Default is \code{365.25}.}

\item{GPD_Bayes}{Logical; indicating whether to use a Bayesian approach to estimate GPD parameters. This involves applying a penalty to the likelihood to aid in the stability of the optimization procedure. Default is \code{TRUE}.}

\item{Method}{Character vector of length one specifying the method of choosing the threshold. \code{"Standard"} (default) chooses the exact threshold specified as either \code{"u"} or \code{"th"}, whereas \code{"Solari"} selects the minimum exceedence of the \code{"Data"} above the user-specified threshold.}

\item{min.RI}{Numeric vector of length one specifying the minimum return period in the return level plot. Default is \code{1}.}

\item{max.RI}{Numeric vector of length one specifying the maximum return period in the return level plot. Default is \code{100}.}

\item{PLOT}{Logical; indicating whether to plot diagnostics. Default is \code{FALSE}.}

\item{xlab_hist}{Character vector of length one. Histogram x-axis label. Default is \code{"Data"}.}

\item{y_lab}{Character vector of length one. Histogram x-axis label. Default is \code{"Data"}.}
}
\value{
List comprising the GPD \code{Threshold}, shape parameter \code{xi} and scale parameters \code{sigma} along with their standard errors \code{sigma.SE} and \code{xi.SE}.
}
\description{
Fit a Generalized Pareto Distribution (GPD) to a declustered dataset.
}
\section{Details}{

For excesses of a variable X over a suitably high threshold u the fitted GPD model is parameterized as follows: \deqn{P( X > x| X > u) = \left[1 + \xi \frac{(x-u)}{\sigma}\right]^{-\frac{1}{\xi}}_{+}}
where \eqn{\xi} and \eqn{\sigma>0} are the shape and scale parameters  of the GPD and \eqn{[y]_{+}=max(y,0)}.
}

\examples{
#Decluster time seires
S20T_decl = Decluster(Data=S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
#Fit GPD
GPD_Fit(Data=S20T_decl$Declustered,
       Data_Full=S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
}
