% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Migpd_Fit.R
\name{Migpd_Fit}
\alias{Migpd_Fit}
\title{Fits Multiple independent generalized Pareto models - Fit}
\usage{
Migpd_Fit(
  Data,
  Data_Full = NA,
  mth,
  mqu,
  penalty = "gaussian",
  maxit = 10000,
  trace = 0,
  verbose = FALSE,
  priorParameters = NULL
)
}
\arguments{
\item{Data}{A data frame with \code{n} columns, each comprising a declustered and if necessary detrended time series to be modelled.}

\item{Data_Full}{A data frame with \code{n} columns, each comprising the original (detrended if necessary) time series to be modelled. Only required if threshold is specified using \code{mqu}.}

\item{mth}{Marginal thresholds, above which generalized Pareto models are fitted. Numeric vector of length \code{n}.}

\item{mqu}{Marginal quantiles, above which generalized Pareto models are fitted. \strong{Only one of \code{mth} and \code{mqu} should be supplied.} Numeric vector of length \code{n}.}

\item{penalty}{See \code{\link{ggplot.migpd}}.}

\item{maxit}{See \code{\link{ggplot.migpd}}.}

\item{trace}{See \code{\link{ggplot.migpd}}.}

\item{verbose}{See \code{\link{ggplot.migpd}}.}

\item{priorParameters}{See \code{\link{ggplot.migpd}}.}
}
\value{
An object of class \code{"migpd"}. There are \code{coef}, \code{print}, \code{plot}, \code{ggplot} and \code{summary} functions available.
}
\description{
Fit multiple independent generalized Pareto models to each column of a data frame. Edited version of the \code{migpd} function in \code{texmex}, to allow for \code{NA}s in a time series.
}
\examples{
#With date as first column
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df, mqu =c(0.99,0.99,0.99))
#Without date as first column
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1], mqu =c(0.99,0.99,0.99))
#Same GPDs fit as above but thresholds given on the original scale
S22.Rainfall.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Rainfall),0.99)
S22.OsWL.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$OsWL),0.99)
S22.GW.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Groundwater),0.99)
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1], mqu =c(S22.Rainfall.Quantile,S22.OsWL.Quantile,S22.GW.Quantile))
}
\seealso{
\code{\link{Decluster}} \code{\link{Detrend}} \code{\link{Dataframe_Combine}}
}
