% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Migpd_Fit.R
\name{Migpd_Fit}
\alias{Migpd_Fit}
\title{Fits Multiple independent generalized Pareto models - Fit}
\usage{
Migpd_Fit(
  Data,
  Data_Full = NA,
  mth,
  mqu,
  penalty = "gaussian",
  maxit = 10000,
  trace = 0,
  verbose = FALSE,
  priorParameters = NULL
)
}
\arguments{
\item{Data}{A data frame with \code{n} columns, each comprising a declustered and if necessary detrended time series to be modelled.}

\item{Data_Full}{A data frame with \code{n} columns, each comprising the original (detrended if necessary) time series to be modelled. Only required if threshold is specified using \code{mqu}.}

\item{mth}{Marginal thresholds, above which generalized Pareto models are fitted. Numeric vector of length \code{n}.}

\item{mqu}{Marginal quantiles, above which generalized Pareto models are fitted. \strong{Only one of \code{mth} and \code{mqu} should be supplied.} Numeric vector of length \code{n}.}

\item{penalty}{How the likelihood should be penalized. Defaults to \code{"gaussian"}.}

\item{maxit}{Numeric vector of length specifying the maximum number of iterations used in the optimization. Default is \code{10000}.}

\item{trace}{Default is \code{0}.}

\item{verbose}{Default is \code{FALSE}.}

\item{priorParameters}{Prior parameters. Only use if \code{penalty = "gaussian"}.}
}
\value{
An object of class \code{"migpd"}. There are \code{coef}, \code{print}, \code{plot}, and \code{summary} functions available.
}
\description{
Fit multiple independent generalized Pareto models to each column of a data frame. Edited version of the \code{migpd} function in \code{texmex}, to allow for \code{NA}s in a time series.
}
\examples{
#With date as first column
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df, Data_Full=S22.Detrend.df, mqu =c(0.99,0.99,0.99))
#Same GPDs fit as above but thresholds given on the original scale
S22.Rainfall.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Rainfall),0.99)
S22.OsWL.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$OsWL),0.99)
S22.GW.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Groundwater),0.99)
S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1],
                   mth =c(S22.Rainfall.Quantile,S22.OsWL.Quantile,S22.GW.Quantile))
}
\seealso{
\code{\link{Decluster}} \code{\link{Detrend}} \code{\link{Dataframe_Combine}}
}
