% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOAAetal2012}
\alias{NOAAetal2012}
\title{2012 NOAA Sea Level Rise Projections for Virginia Key}
\format{
A data frame containing the projections with columns as follows:
\describe{
\item{Year}{Year of Projection (YYYY)}
\item{Low}{Projections for the low emission scenario (m)}
\item{Int.Low}{Projections for the intermediate low emission scenario (m)}
\item{Int.High}{Projections for the intermediate high emission scenario (m)}
\item{High}{Projections for the high emission scenario (m)}
}
}
\source{
NOAA National Ocean Service. Report available at:
\url{https://repository.library.noaa.gov/view/noaa/11124}
}
\usage{
data(NOAAetal2012)
}
\description{
Sea level rise projections for Virginia Key under four emission scenarios from NOAA Technical Report OAR CPO-1 (2012).
}
\keyword{datasets}
