% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOAAetal2017}
\alias{NOAAetal2017}
\title{2017 NOAA Sea Level Rise Projections for Virginia Key}
\format{
A data frame containing the projections with columns as follows:
\describe{
\item{Year}{Year of Projection (YYYY)}
\item{VLM}{Virtical land motion (m)}
\item{Low}{Projections for the low emission scenario (m)}
\item{Int.Low}{Projections for the intermediate low emission scenario (m)}
\item{Intermediate}{Projections for the intermediate emission scenario (m)}
\item{Int.High}{Projections for the intermediate high emission scenario (m)}
\item{High}{Projections for the high emission scenario (m)}
\item{Extreme}{Projections for the extreme emission scenario (m)}
}
}
\source{
NOAA National Ocean Service. Report available at: \url{https://tidesandcurrents.noaa.gov/publications/techrpt83_Global_and_Regional_SLR_Scenarios_for_the_US_final.pdf}
}
\usage{
data(NOAAetal2017)
}
\description{
Sea level rise projections for Virginia Key derived from NOAA Technical Report NOS CO-OPS 083 (2017). The report defines six global mean sea level rise scenarios, which are regionally downscaled on a 1-degree grid covering U.S. coastlines.
}
\keyword{datasets}
