% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{S13.Detrend.df}
\alias{S13.Detrend.df}
\title{Measurement from Control Structure S-13}
\format{
A data frame with 281,846 rows and 3 variables:
\describe{
\item{Date_Time}{Time of observation (YYYY-MM-DD HH:MM:SS)}
\item{Rainfall}{Rainfall totals (inches)}
\item{OsWL}{Ocean-side water levels totals (ft NGVD29)}
}
}
\source{
South Florida Water Management District, via
\url{https://www.sfwmd.gov/science-data/dbhydro/}
}
\usage{
data(S13.Detrend.df)
}
\description{
Hourly time series of rainfall totals and ocean-side water levels (tailwater levels) measured at control structure S-13. The O-sWLs are detrended using a 3-month moving window.
}
\keyword{datasets}
