% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vine_Copula_Fit.R
\name{Vine_Copula_Fit}
\alias{Vine_Copula_Fit}
\title{C and D-vine Copula - Fitting}
\usage{
Vine_Copula_Fit(Data, FamilySet = NA, Type = "DVine",
  SelCrit = "AIC", Indeptest = FALSE, Level = 0.05)
}
\arguments{
\item{Data}{Dataframe containing \code{n} at least partially concurrent time series. First column may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}

\item{FamilySet}{Integer vector whuch must include at least one pair-copula family that allows for positive and one that allows for negative dependence. If \code{familyset = NA} (default), selection among all possible families is performed. The coding of pair-copula families is shown below. See help file of the \code{CDVineSim} function to find out copula represented by integers \code{0-40}.}

\item{Type}{Type of the vine model:\itemize{
\item 1 or "CVine" = C-vine
\item 2 or "DVine" = D-vine
}}

\item{SelCrit}{Character vector specifying the criterion for choosing among the competing pair-copula. Possible choices: \code{"AIC"} (default) or \code{"BIC"}.}

\item{Indeptest}{Logical; whether a hypothesis test for the independence of \code{u1} and \code{u2} is performed before bivariate copula selection (default: \code{Indeptest = FALSE}; cp. BiCopIndTest). The independence copula is chosen for a (conditional) pair if the null hypothesis of independence cannot be rejected.}

\item{level}{Numeric; significance level of the independence test (default: level = 0.05).}
}
\value{
List comprising the pair-copula families composing the C- or D-vine copula \code{Family}, its parameters \code{Par} and \code{Par2} as well as whether it is a C or D-vine \code{Type}.
}
\description{
Fit either a C- or D-vine copula model. Function is a repackaging of the \code{CDVineCopSelect} function in the \code{CDVine} package.
}
\examples{
S22.Vine<-Vine_Copula_Fit(Data=S22.Detrend.df, FamilySet=NA, Type="DVine", SelCrit="AIC",Indeptest=FALSE, Level=0.05)
}
\seealso{
\code{\link{Dataframe_Combine}} \code{\link{Vine_Copula_Sim}} \code{\link{CDVineCopSelect}}  \code{\link{BiCopSelect}}
}
