% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vine_Copula_Sim.R
\name{Vine_Copula_Sim}
\alias{Vine_Copula_Sim}
\title{C and D-vine Copula - Simulation}
\usage{
Vine_Copula_Sim(Data, Vine_Model, Marginals, mu = 365.25, N = 10000)
}
\arguments{
\item{Data}{Data frame containing \code{n} at least partially concurrent time series. First column may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}

\item{Vine_Model}{An \code{RVineMatrix} object i.e., output of \code{Vine_Copula_Fit} specifying the structure and copula families composing the vine copula.}

\item{Marginals}{An \code{migpd} object containing the d-independent generalized Pareto models.}

\item{mu}{(average) Number of events per year. Numeric vector of length one. Default is 365.25, daily data.}

\item{N}{Number of years worth of extremes to be simulated. Numeric vector of length one. Default 10,000 (years).}
}
\value{
List comprising an integer vector specifying the pair-copula families composing the C- or D-vine copula \code{Vine_family}, its parameters \code{Vine_par} and \code{Vine_par2} and type of regular vine \code{Vine_Type}. In addition, data frames of the simulated observations: \code{u.Sim} on the transformed \code{$[0,1]^n$} and \code{x.Sim} the original scales.
}
\description{
Simulating from specified C- and D-vine copula models. Function is a repackaging of the \code{RVineMatrix} and \code{RVineMatrix} functions from the \code{VineCopula} package into a single function.
}
\examples{
#Fitting GPD to independent cluster maxima
S20.Migpd<-Migpd_Fit(Data=S20.Detrend.Declustered.df,
                     Data_Full=S20.Detrend.df,
                     mqu=c(0.975,0.975,0.9676))
#Fitting vine copula
S20.Vine<-Vine_Copula_Fit(Data=S20.Detrend.df)
#Simulating from fitted copula
S20.Vine.Sim<-Vine_Copula_Sim(Data=S20.Detrend.df,Vine_Model=S20.Vine,
                              Marginals=S20.Migpd,N=10)
#Plotting observed (black) and simulated (red) values
S20.Pairs.Plot.Data<-data.frame(rbind(na.omit(S20.Detrend.df[,-1]),S22.Vine.Sim$x.Sim),
                                c(rep("Observation",nrow(na.omit(S20.Detrend.df))),
                                rep("Simulation",nrow(S20.Vine.Sim$x.Sim))))
colnames(S20.Pairs.Plot.Data)<-c(names(S20.Detrend.df)[-1],"Type")
pairs(S20.Pairs.Plot.Data[,1:3],
      col=ifelse(S20.Pairs.Plot.Data$Type=="Observation","Black","Red"),
      upper.panel=NULL)
}
\seealso{
\code{\link{Vine_Copula_Fit}}
}
